Rest_api/requirements.txt
paho-mqtt==1.6.1
pymongo==4.4.1
python-dotenv==1.0.0
uvicorn==0.23.2
fastapi==0.101.1
